src/pyproject.toml
[tool.poetry]
name = "cs2-battle-bot-api"
version = "0.0.5"
description = ""
authors = ["Adrian Ciolek <ciolek.adrian@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
httpx = "^0.26.0"
psycopg2-binary = "^2.9.9"
django-prefix-id = "^1.0.0"
steam = "^1.4.4"
django-redis = "^5.4.0"
djangorestframework = "^3.14.0"
djangorestframework-api-key = "==2.*"
rcon = "^2.4.6"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
